9610 qede: error: left shift of negative value Contributed by: Anand Khoje <Anand.Khoje@cavium.com> Reviewed by: Andy Fiddaman <af@citrus-it.net> Approved by: Dan McDonald <danmcd@joyent.com>

+++ usr/src/uts/common/io/qede/579xx/drivers/ecore/ecore.h
@@ -111,7 +111,7 @@ enum ecore_nvm_cmd {
 
 #define SET_FIELD(value, name, flag)					\
 do {									\
-	(value) &= ~(name##_MASK << name##_SHIFT);			\
+	(value) &= ~((u64)name##_MASK << (u64)name##_SHIFT);		\
 	(value) |= ((((u64)flag) & (u64)name##_MASK) << (name##_SHIFT));\
 } while (0)
